requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
mlxtend
mysql-connector-python
logging